[Refactor] Extract lake connector into a clean module
