+++ be/src/connector/connector_registry_init.cpp
@@ -18,7 +18,6 @@
 #include "connector/builtin_connector_registry.h"
 #include "connector/connector_registry.h"
 #include "connector/file/file_connector.h"
-#include "connector/lake_connector.h"
 
 namespace starrocks::connector {
 
@@ -36,7 +35,6 @@ void install_if_absent(ConnectorRegistry* registry, const std::string& name) {
 Status install_builtin_connectors(ConnectorRegistry* registry) {
     DCHECK(registry != nullptr);
     install_if_absent<FileConnector>(registry, Connector::FILE);
-    install_if_absent<LakeConnector>(registry, Connector::LAKE);
     return Status::OK();
 }
