+++ be/src/connector/lake/lake_connector.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "connector/lake_connector.h"
+#include "connector/lake/lake_connector.h"
 
 #include <vector>
