+++ be/test/connector/CMakeLists.txt
@@ -121,6 +121,25 @@ add_executable(connector_iceberg_test
 target_link_libraries(connector_iceberg_test ${CONNECTOR_ICEBERG_TEST_LINK_LIBS})
 set_target_properties(connector_iceberg_test PROPERTIES COMPILE_FLAGS "-fno-access-control")
 
+set(CONNECTOR_LAKE_TEST_LINK_LIBS
+    ConnectorLake
+    ${CONNECTOR_COMMON_TEST_DEPENDENCIES}
+    ${STARROCKS_STATIC_LINK_LIBS}
+    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
+    ${WL_START_GROUP}
+    gmock
+    gtest
+    ${WL_END_GROUP}
+)
+
+add_executable(connector_lake_test
+        ${BE_TEST_MACOS_LINK_STUBS}
+        ${BE_TEST_JNI_ENV_STUB}
+        lake/lake_global_late_materialization_context_test.cpp
+)
+target_link_libraries(connector_lake_test ${CONNECTOR_LAKE_TEST_LINK_LIBS} gtest_main)
+set_target_properties(connector_lake_test PROPERTIES COMPILE_FLAGS "-fno-access-control")
+
 set(CONNECTOR_CACHE_STATS_TEST_LINK_LIBS
     ConnectorCacheStats
     ${STARROCKS_DEPENDENCIES}
