+++ be/test/module/CMakeLists.txt
@@ -12,7 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_executable(module_bootstrap_test module_bootstrap_test.cpp)
+set(MODULE_BOOTSTRAP_TEST_EXTRA_SOURCES)
+if (APPLE)
+    list(APPEND MODULE_BOOTSTRAP_TEST_EXTRA_SOURCES ../base/macos_link_stubs.cpp)
+endif()
+
+add_executable(module_bootstrap_test
+        ${MODULE_BOOTSTRAP_TEST_EXTRA_SOURCES}
+        module_bootstrap_test.cpp)
 target_link_be_test_cxa_throw_wrap(module_bootstrap_test)
 set(MODULE_BOOTSTRAP_TEST_LINK_LIBS ${TEST_LINK_LIBS})
 list(REMOVE_ITEM MODULE_BOOTSTRAP_TEST_LINK_LIBS arrow_testing)
