+++ be/test/storage/lake/lake_scan_node_test.cpp
@@ -22,7 +22,7 @@
 #include "column/vectorized_fwd.h"
 #include "common/config_scan_io_fwd.h"
 #include "common/logging.h"
-#include "connector/lake_connector.h"
+#include "connector/lake/lake_connector.h"
 #include "exec/connector_scan_node.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
 #include "storage/chunk_helper.h"
